+++ src/plugins/vis_type_tagcloud/public/components/tag_cloud.test.js
@@ -32,9 +32,12 @@ import _ from 'lodash';
 import d3 from 'd3';
 import 'jest-canvas-mock';
 
-import { fromNode, delay } from 'bluebird';
 import { TagCloud } from './tag_cloud';
 import { setHTMLElementOffset, setSVGElementGetBBox } from '../../../../test_utils/public';
+
+function delay(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
 
 describe('tag cloud tests', () => {
   let SVGElementGetBBoxSpyInstance;
@@ -149,7 +152,11 @@ describe('tag cloud tests', () => {
         tagCloud = new TagCloud(domNode, colorScale);
         tagCloud.setData(currentTest.data);
         tagCloud.setOptions(currentTest.options);
-        await fromNode((cb) => tagCloud.once('renderComplete', cb));
+        await new Promise((resolve) => {
+          tagCloud.once('renderComplete', () => {
+            resolve();
+          });
+        });
       });
 
       afterEach(teardownDOM);
@@ -183,7 +190,11 @@ describe('tag cloud tests', () => {
         //this timeout modifies the settings before the cloud is rendered.
         //the cloud needs to use the correct options
         setTimeout(() => tagCloud.setOptions(logScaleTest.options), timeout);
-        await fromNode((cb) => tagCloud.once('renderComplete', cb));
+        await new Promise((resolve) => {
+          tagCloud.once('renderComplete', () => {
+            resolve();
+          });
+        });
       });
 
       afterEach(teardownDOM);
@@ -211,7 +222,11 @@ describe('tag cloud tests', () => {
       tagCloud.setData(baseTest.data);
       tagCloud.setOptions(baseTest.options);
       tagCloud.setOptions(logScaleTest.options);
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -238,7 +253,11 @@ describe('tag cloud tests', () => {
       tagCloud.setOptions(baseTest.options);
       tagCloud.setData(trimDataTest.data);
 
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -324,7 +343,11 @@ describe('tag cloud tests', () => {
         tagCloud.setData(baseTest.data);
         tagCloud.setOptions(baseTest.options);
       }, 200);
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -349,7 +372,11 @@ describe('tag cloud tests', () => {
       tagCloud = new TagCloud(domNode, colorScale);
       tagCloud.setData(baseTest.data);
       tagCloud.setOptions(baseTest.options);
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -371,12 +398,20 @@ describe('tag cloud tests', () => {
       tagCloud = new TagCloud(domNode, colorScale);
       tagCloud.setData(baseTest.data);
       tagCloud.setOptions(baseTest.options);
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
 
       //make bigger
       tagCloud._size = [600, 600];
       tagCloud.resize();
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -394,12 +429,20 @@ describe('tag cloud tests', () => {
       tagCloud = new TagCloud(domNode, colorScale);
       tagCloud.setData(baseTest.data);
       tagCloud.setOptions(baseTest.options);
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
 
       //make smaller
       tagCloud._size = [];
       tagCloud.resize();
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
     });
 
     afterEach(teardownDOM);
@@ -417,7 +460,11 @@ describe('tag cloud tests', () => {
       tagCloud.setData(baseTest.data);
       tagCloud.setOptions(baseTest.options);
 
-      await fromNode((cb) => tagCloud.once('renderComplete', cb));
+      await new Promise((resolve) => {
+        tagCloud.once('renderComplete', () => {
+          resolve();
+        });
+      });
 
       expect(domNode.innerHTML).toMatchSnapshot();
     });